requirements.txt
httpx
mcp
tavily-python>=0.5.0
uvicorn